pyproject.toml
[tool.poetry]
name = "fastapidbtproject"
version = "0.1.0"
description = "DBT Skills diary cards service."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"


[tool.poetry.group.diary-ms.dependencies]
fastapi = "^0.115.3"
fastapi-cli = "^0.0.5"
httpx = "^0.27.2"
dishka = "^1.8.0"
alembic = "^1.14.0"
greenlet = "^3.1.1"
asyncpg = "^0.30.0"
orjson = "^3.10.12"
sqlalchemy = {extras = ["asycio"], version = "^2.0.36"}
pyjwt = "^2.10.1"
taskiq-redis = "^1.0.2"
faststream = {extras = ["kafka"], version = "^0.5.34"}
taskiq-fastapi = "^0.3.3"
boto3 = "1.35.99"
reportlab = "^4.3.1"
opentelemetry-api = "^1.30.0"
opentelemetry-sdk = "^1.30.0"
opentelemetry-instrumentation-fastapi = "^0.51b0"
opentelemetry-exporter-otlp = "^1.30.0"
opentelemetry-instrumentation-logging = "^0.51b0"
opentelemetry-distro = "^0.51b0"
prometheus-client = "^0.21.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.4"
mypy = "^1.13.0"
pre-commit = "^4.0.1"
pytest-asyncio = "^0.24.0"
redis-cli = "^1.0.1"



[tool.poetry.group.user-ms.dependencies]
django = "^5.1.4"
djangorestframework = "^3.15.2"
djoser = "^2.3.1"
djangorestframework-simplejwt = "^5.3.1"
social-auth-app-django = "^5.4.2"
django-environ = "^0.11.2"
psycopg = {extras = ["binary"], version = "^3.2.3"}
drf-spectacular = "^0.28.0"


[tool.poetry.group.diary-bot.dependencies]
aiogram = "^3.18.0"
aiogram-dialog = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "migrations", "django", "reportlab", "src.user_ms"]
plugins = ['pydantic.mypy']


[[tool.mypy.overrides]]
module = [
    "taskiq_faststream",
    "taskiq_fastapi",
    "botocore.config",
    "botocore.client",
    "boto3",
    "reportlab.*"
]
ignore_missing_imports = true

[tool.pyright]
pythonVersion = "3.13"

[tool.ruff]
target-version = "py313"
exclude = ["migrations"]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.pytest.ini_options]
pythonpath = ".src"
asyncio_mode = "auto"
python_files = "*_test.py *_tests.py test_*.py"

testpaths = [
    "tests",
    "integration",
]

